updated to tiltfive 1.3.0 (still very much a work in progress)

+++ glarses/io/file.cpp
@@ -6,7 +6,7 @@ namespace io {
 	std::string read_text_file(const std::filesystem::path& p) {
 		std::string result;
 
-		size_t num_bytes = std::filesystem::file_size(p);
+		size_t num_bytes = static_cast<size_t>(std::filesystem::file_size(p));
 		
 		result.resize(num_bytes);
